+++ lib/cli/commands/index.js
@@ -1,2 +1,2 @@
-export { default as interactive } from './interactive.js';
+export { default as interactive } from './interactive-command.js';
 export { default as submenu } from './create-submenu-patch-command.js';
