+++ lib/cli/commands/interactive-command.js
@@ -1,7 +1,7 @@
 // # interactive.js
-import * as commands from '../commands/index.js';
-import * as prompts from '../prompts.js';
-import * as flows from '../flows/index.js';
+import * as commands from '#cli/commands';
+import * as prompts from '#cli/prompts';
+import * as flows from '#cli/flows';
 
 // The default command when the program is used without any options. In this 
 // case we fire up an inquirer session to ask the user what they want to do.
