+++ lib/cli/flows/index.js
@@ -1,1 +1,1 @@
-export { default as submenu } from './submenu.js';
+export { default as submenu } from './submenu-flow.js';
