+++ lib/cli/flows/submenu-flow.js
@@ -2,7 +2,7 @@
 import chalk from 'chalk';
 import path from 'node:path';
 import fs from 'node:fs';
-import * as prompts from '../prompts.js';
+import * as prompts from '#cli/prompts';
 
 // # submenu()
 // Contains the interactive flow for getting the options to pass to the submenu
