+++ lib/cli/prompts/index.js
@@ -5,7 +5,7 @@ export { default as confirm } from '@inquirer/confirm';
 export { default as number } from '@inquirer/number';
 export { default as select } from '@inquirer/select';
 export { default as checkbox, Separator } from '@inquirer/checkbox';
-export { default as nestedList } from './prompts/nested-list.js';
+export { default as nestedList } from './nested-list.js';
 
 // # hex(opts)
 // Helper function for requesting a single hexadecimal number as input, with
