+++ lib/cli/setup-cli.js
@@ -4,8 +4,8 @@ import chalk from 'chalk';
 import tar from 'tar';
 import ora from 'ora';
 import { program as commander, Command } from 'commander';
-import * as commands from './commands/index.js';
-import * as prompts from './prompts.js';
+import * as commands from '#cli/commands';
+import * as prompts from '#cli/prompts';
 import inquirer from 'inquirer';
 import { DBPF, Savegame, FileType, ZoneType } from 'sc4/core';
 import * as api from 'sc4/api';
